Switched BSP to SYG_MDK
Switched BSP to SYG_MDK since I needed to make a new pack file for the SYG_MDK for a customer anyway.

The only difference between this pack file and the RoverDemo pack file was the roverdemo.pincfg.
The actual BSP files are exactly the same since the QSPI, SDRAM, and LEDs are all on-board the SOM module.
This repository preserves the pincfg file as well as the original SSP 1.2.0 pack file if some reason exists to re-create the pack file for newer SSP versions.